requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.1
billiard==4.2.4
boto3==1.42.57
botocore==1.42.57
bytecode==0.17.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
Django==5.2.11
django-storages==1.14.6
djangorestframework==3.16.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==25.1.0
httplib2==0.31.2
idna==3.11
jmespath==1.1.0
kombu==5.6.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.2.0
Redy==0.2.10
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
